+++ Source/CesiumRuntime/Private/CesiumGeoreference.cpp
@@ -4,7 +4,9 @@
 #include "Camera/PlayerCameraManager.h"
 #include "CesiumActors.h"
 #include "CesiumCommon.h"
+#include "CesiumCustomVersion.h"
 #include "CesiumGeospatial/Cartographic.h"
+#include "CesiumOriginShiftComponent.h"
 #include "CesiumRuntime.h"
 #include "CesiumSubLevelComponent.h"
 #include "CesiumSubLevelSwitcherComponent.h"
@@ -36,6 +38,7 @@
 #include "DrawDebugHelpers.h"
 #include "Editor.h"
 #include "EditorViewportClient.h"
+#include "GameFramework/Pawn.h"
 #include "Slate/SceneViewport.h"
 #endif
 
@@ -206,11 +209,11 @@ void ACesiumGeoreference::SetScale(double NewValue) {
 }
 
 APlayerCameraManager* ACesiumGeoreference::GetSubLevelCamera() const {
-  return this->SubLevelCamera;
+  return this->SubLevelCamera_DEPRECATED;
 }
 
 void ACesiumGeoreference::SetSubLevelCamera(APlayerCameraManager* NewValue) {
-  this->SubLevelCamera = NewValue;
+  this->SubLevelCamera_DEPRECATED = NewValue;
 }
 
 #if WITH_EDITOR
@@ -368,7 +371,7 @@ void ACesiumGeoreference::_showSubLevelLoadRadii() const {
   }
 
   for (const auto& pLevelWeak :
-       this->SubLevelSwitcher->GetRegisteredSubLevels()) {
+       this->SubLevelSwitcher->GetRegisteredSubLevelsWeak()) {
     ALevelInstance* pLevel = pLevelWeak.Get();
     if (!IsValid(pLevel))
       continue;
@@ -401,14 +404,12 @@ void ACesiumGeoreference::Tick(float DeltaTime) {
 #if WITH_EDITOR
   _showSubLevelLoadRadii();
 #endif
-
-  if (this->_shouldManageSubLevels()) {
-    _updateSublevelState();
-  }
 }
 
 void ACesiumGeoreference::Serialize(FArchive& Ar) {
   Super::Serialize(Ar);
+
+  Ar.UsingCustomVersion(FCesiumCustomVersion::GUID);
 
   // Recompute derived values on load.
   if (Ar.IsLoading()) {
@@ -428,22 +429,6 @@ void ACesiumGeoreference::BeginPlay() {
         Warning,
         TEXT("CesiumGeoreference does not have a World in BeginPlay."));
     return;
-  }
-
-  if (!this->SubLevelCamera) {
-    // Find the first player's camera manager
-    APlayerController* pPlayerController = pWorld->GetFirstPlayerController();
-    if (pPlayerController) {
-      this->SubLevelCamera = pPlayerController->PlayerCameraManager;
-    }
-
-    if (!this->SubLevelCamera) {
-      UE_LOG(
-          LogCesium,
-          Warning,
-          TEXT(
-              "CesiumGeoreference could not find a FirstPlayerController or a corresponding PlayerCameraManager."));
-    }
   }
 
   UpdateGeoreference();
@@ -471,6 +456,94 @@ void ACesiumGeoreference::PostLoad() {
       IsValid(this->GetWorld()->WorldComposition) &&
       !this->GetWorld()->IsGameWorld()) {
     this->_createSubLevelsFromWorldComposition();
+  }
+
+  const int32 CesiumVersion =
+      this->GetLinkerCustomVersion(FCesiumCustomVersion::GUID);
+
+  if (CesiumVersion < FCesiumCustomVersion::OriginShiftComponent &&
+      !this->SubLevelSwitcher->GetRegisteredSubLevelsWeak().IsEmpty()) {
+    // In previous versions, the CesiumGeoreference managed origin shifting
+    // based on a SubLevelCamera, which defaulted to the PlayerCameraManager of
+    // the World's `GetFirstPlayerController()`.
+
+    // Backward compatibility for this is tricky, but we can make a decent
+    // attempt that will work in a lot of cases. And this is just an unfortunate
+    // v2.0 breakage for any remaining cases.
+
+    AActor* SubLevelActor = nullptr;
+
+    if (this->SubLevelCamera_DEPRECATED) {
+      // An explicit SubLevelCamera_DEPRECATED is specified. If it has a target
+      // Actor, attach a CesiumOriginShiftComponent to that Actor.
+      SubLevelActor = this->SubLevelCamera_DEPRECATED->ViewTarget.Target.Get();
+
+      if (!SubLevelActor) {
+        UE_LOG(
+            LogCesium,
+            Warning,
+            TEXT("An explicit SubLevelCamera was specified on this "
+                 "CesiumGeoreference, but its ViewTarget is not a valid "
+                 "Actor, so a CesiumOriginShiftComponent could not be added "
+                 "automatically. You must manually add a "
+                 "CesiumOriginShiftComponent to the Actor whose position "
+                 "should be used to control sub-level switching."));
+      }
+    } else {
+      // No explicit SubLevelCamera_DEPRECATED, so try to find a Pawn set to
+      // auto-possess player 0.
+      for (TActorIterator<APawn> it(
+               GetWorld(),
+               APawn::StaticClass(),
+               EActorIteratorFlags::SkipPendingKill);
+           it;
+           ++it) {
+        if (it->AutoPossessPlayer == EAutoReceiveInput::Player0) {
+          SubLevelActor = *it;
+          break;
+        }
+      }
+
+      if (!SubLevelActor) {
+        UE_LOG(
+            LogCesium,
+            Warning,
+            TEXT(
+                "Could not find a Pawn in the level set to auto-possess player "
+                "0, so a CesiumOriginShiftComponent could not be added "
+                "automatically. You must manually add a "
+                "CesiumOriginShiftComponent to the Actor whose position "
+                "should be used to control sub-level switching."));
+      }
+    }
+
+    if (SubLevelActor) {
+      // If this is a Blueprint object, like DynamicPawn, its construction
+      // scripts may not have been run yet at this point. Doing so might cause
+      // an origin shift component to be added. So we force it to happen here so
+      // that we don't end up adding a duplicate CesiumOriginShiftComponent.
+      SubLevelActor->RerunConstructionScripts();
+      if (SubLevelActor->FindComponentByClass<UCesiumOriginShiftComponent>() ==
+          nullptr) {
+
+        UCesiumOriginShiftComponent* OriginShift =
+            Cast<UCesiumOriginShiftComponent>(
+                SubLevelActor->AddComponentByClass(
+                    UCesiumOriginShiftComponent::StaticClass(),
+                    false,
+                    FTransform::Identity,
+                    false));
+        OriginShift->SetFlags(RF_Transactional);
+        SubLevelActor->AddInstanceComponent(OriginShift);
+
+        UE_LOG(
+            LogCesium,
+            Warning,
+            TEXT("Added CesiumOriginShiftComponent to %s in order to preserve "
+                 "backward compatibility for sub-level switching."),
+            *SubLevelActor->GetName());
+      }
+    }
   }
 #endif // WITH_EDITOR
 }
@@ -519,7 +592,7 @@ void ACesiumGeoreference::_createSubLevelsFromWorldComposition() {
   }
 
   if (this->CesiumSubLevels_DEPRECATED.IsEmpty() ||
-      !this->_shouldManageSubLevels())
+      !this->GetLevel()->IsPersistentLevel())
     return;
 
   this->Modify();
@@ -582,8 +655,8 @@ void ACesiumGeoreference::_createSubLevelsFromWorldComposition() {
     pLevelComponent->SetEnabled(pFound->Enabled);
     pLevelComponent->SetLoadRadius(pFound->LoadRadius);
 
-    // But if the georeference origin is close to this sub-level's origin, make
-    // this the active sub-level.
+    // But if the georeference origin is close to this sub-level's origin,
+    // make this the active sub-level.
     if (FMath::IsNearlyEqual(
             this->OriginLongitude,
             pFound->LevelLongitude,
@@ -599,7 +672,7 @@ void ACesiumGeoreference::_createSubLevelsFromWorldComposition() {
     pLevelInstance->LoadLevelInstance();
   }
 
-  this->SubLevelSwitcher->SetTarget(pActiveSubLevel);
+  this->SubLevelSwitcher->SetTargetSubLevel(pActiveSubLevel);
 
   this->CesiumSubLevels_DEPRECATED.Empty();
 
@@ -648,10 +721,10 @@ void ACesiumGeoreference::UpdateGeoreference() {
 
   // If we're in a sub-level, update its origin as well.
   UCesiumSubLevelSwitcherComponent* pSwitcher = this->SubLevelSwitcher;
-  if (IsValid(pSwitcher) && pSwitcher->GetTarget() != nullptr) {
-    if (pSwitcher->GetTarget() == pSwitcher->GetCurrent() ||
-        pSwitcher->GetCurrent() == nullptr) {
-      ALevelInstance* pTarget = pSwitcher->GetTarget();
+  if (IsValid(pSwitcher) && pSwitcher->GetTargetSubLevel() != nullptr) {
+    if (pSwitcher->GetTargetSubLevel() == pSwitcher->GetCurrentSubLevel() ||
+        pSwitcher->GetCurrentSubLevel() == nullptr) {
+      ALevelInstance* pTarget = pSwitcher->GetTargetSubLevel();
       UCesiumSubLevelComponent* pComponent =
           pTarget->FindComponentByClass<UCesiumSubLevelComponent>();
       if (IsValid(pComponent)) {
@@ -683,69 +756,6 @@ GeoTransforms ACesiumGeoreference::GetGeoTransforms() const noexcept {
 FName ACesiumGeoreference::DEFAULT_GEOREFERENCE_TAG =
     FName("DEFAULT_GEOREFERENCE");
 
-bool ACesiumGeoreference::_updateSublevelState() {
-  const TArray<TWeakObjectPtr<ALevelInstance>>& sublevels =
-      this->SubLevelSwitcher->GetRegisteredSubLevels();
-
-  if (sublevels.Num() == 0) {
-    // If we don't have any known sub-levels, bail quickly to save ourselves a
-    // little work.
-    return false;
-  }
-
-  if (!IsValid(this->SubLevelCamera)) {
-    return false;
-  }
-
-  UWorld* pWorld = this->GetWorld();
-  const FIntVector& originLocation = pWorld->OriginLocation;
-  const FMinimalViewInfo& pov = this->SubLevelCamera->ViewTarget.POV;
-  const FVector& cameraLocation = pov.Location;
-
-  // Transform camera from World to Georeference local frame
-  FVector cameraRelativeToGeoreference =
-      this->GetActorTransform().TransformPosition(cameraLocation);
-
-  // Transform camera from Georeference local frame to ECEF
-  FVector cameraEcef = this->TransformUnrealPositionToEarthCenteredEarthFixed(
-      cameraRelativeToGeoreference);
-
-  ALevelInstance* pClosestActiveLevel = nullptr;
-  double closestLevelDistance = std::numeric_limits<double>::max();
-
-  for (int32 i = 0; i < sublevels.Num(); ++i) {
-    ALevelInstance* pCurrent = sublevels[i].Get();
-    if (!IsValid(pCurrent))
-      continue;
-
-    UCesiumSubLevelComponent* pComponent =
-        pCurrent->FindComponentByClass<UCesiumSubLevelComponent>();
-    if (!IsValid(pComponent))
-      continue;
-
-    if (!pComponent->GetEnabled())
-      continue;
-
-    FVector levelEcef =
-        UCesiumWgs84Ellipsoid::LongitudeLatitudeHeightToEarthCenteredEarthFixed(
-            FVector(
-                pComponent->GetOriginLongitude(),
-                pComponent->GetOriginLatitude(),
-                pComponent->GetOriginHeight()));
-
-    double levelDistance = FVector::Distance(levelEcef, cameraEcef);
-    if (levelDistance < pComponent->GetLoadRadius() &&
-        levelDistance < closestLevelDistance) {
-      pClosestActiveLevel = pCurrent;
-      closestLevelDistance = levelDistance;
-    }
-  }
-
-  this->SubLevelSwitcher->SetTarget(pClosestActiveLevel);
-
-  return pClosestActiveLevel != nullptr;
-}
-
 void ACesiumGeoreference::_updateCoordinateSystem() {
   if (this->OriginPlacement == EOriginPlacement::CartographicOrigin) {
     FVector origin = this->GetOriginLongitudeLatitudeHeight();
@@ -765,9 +775,4 @@ void ACesiumGeoreference::_updateCoordinateSystem() {
         glm::dvec4(0.0, 0.0, 0.0, 1.0));
     this->_coordinateSystem = LocalHorizontalCoordinateSystem(localToEcef);
   }
-}
-
-bool ACesiumGeoreference::_shouldManageSubLevels() const {
-  // Only a Georeference in the PersistentLevel should manage sub-levels.
-  return this->GetLevel()->IsPersistentLevel();
 }