+++ Source/CesiumRuntime/Private/CesiumGlobeAnchorComponent.cpp
@@ -42,12 +42,9 @@
 // * Ignores `AdjustOrientationForGlobeWhenMoving` because the globe position is
 // not changing.
 //
-// ## Origin Rebased
+// ## OriginLocation Changed
 //
-// * Detected by a call to `ApplyWorldOffset`.
-// * Updates the Actor transform from the existing ECEF transform.
-// * Ignores `AdjustOrientationForGlobeWhenMoving` because the globe position is
-// not changing.
+// * Handled by Unreal's normal `ApplyWorldOffset` mechanism.
 
 namespace {
 
@@ -611,7 +608,7 @@ CesiumGeospatial::GlobeAnchor UCesiumGlobeAnchorComponent::
   assert(pGeoreference != nullptr);
 
   const CesiumGeospatial::LocalHorizontalCoordinateSystem& local =
-      pGeoreference->getCoordinateSystem();
+      pGeoreference->GetCoordinateSystem();
 
   glm::dmat4 newModelToLocal =
       VecMath::createMatrix4D(newRelativeTransform.ToMatrixWithScale());
@@ -662,7 +659,7 @@ void UCesiumGlobeAnchorComponent::_updateFromNativeGlobeAnchor(
   ACesiumGeoreference* pGeoreference = this->ResolveGeoreference();
   if (IsValid(pGeoreference)) {
     glm::dmat4 anchorToLocal = nativeAnchor.getAnchorToLocalTransform(
-        pGeoreference->getCoordinateSystem());
+        pGeoreference->GetCoordinateSystem());
 
     this->_setCurrentRelativeTransform(
         FTransform(VecMath::createMatrix(anchorToLocal)));